godot_media/requirement.txt
Django
psycopg2_binary
django_crispy_forms
djangorestframework
